pyproject.toml
[project]
name = "sqlalchemy_upsert"
version = "0.1.0"
description = "A project for upserting data from CSV files to PostgreSQL and dumping tables from MSSQL to CSV."
authors = [{name = "Joao Brito", email = "machado000@gmail.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "pandas",
    "SQLAlchemy",
    "psycopg2-binary",
    "pyodbc",
    "python-dotenv",
    "cleantext",
    "tqdm",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "sqlalchemy_upsert", from = "src" }]

[tool.poetry.group.dev.dependencies]
pytest = "*"
pandas = "*"
SQLAlchemy = "*"
testcontainers = "*"
tqdm = "*"
